Use EqualsVerifier to test equals methods.
